integraded contour digit code

+++ streamlit_app.py
@@ -172,7 +172,7 @@ st.write("""
 def predict_digit_from_canvas(canvas_data, num_samples):
     if canvas_data is not None:
         # Preprocessing
-        img = grab_digits_from_canvas(canvas_data.astype('float32'))
+        img = grab_digits_from_canvas(canvas_data)
 
         # Prediction
         # pred = model.predict(img, batch_size=num_samples)  # Assume model.predict handles BNN sampling
